requirements/base.txt
argon2-cffi==19.1.0
asn1crypto==0.24.0
awesome-slugify==1.6.5
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
click==7.0
coreapi==2.0.8
croniter==0.3.30
cryptography==2.7
django-crispy-forms==1.7.2
django-environ==0.4.5
django-model-utils==3.1.2
django-redis==4.10.0
django-rq-scheduler==1.1.3
django-rq==2.0
django-slds==0.1.2
django==2.2.2
djangorestframework==3.9.4
github3.py==1.3.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
honcho==1.0.1
httplib2==0.13.0
idna==2.8
jwcrypto==0.6.0
psycopg2-binary==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.1
pyyaml==5.1.1
redis==3.2.1
regex==2019.6.8
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.25.3
wheel==0.33.4
whitenoise==4.1.2
